requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
RPi.GPIO==0.7.1
picamera2==0.3.12
Pillow==10.0.1
